+++ packages/core/src/suggest/SuggestControl.ts
@@ -152,6 +152,7 @@ export default class SuggestControl<V, O extends SuggestOptionModel<V>> extends 
   private cancel: () => void = () => {
     this.setState({
       show: false,
+      focused: false,
     })
     if (this.props.onCancel) {
       this.props.onCancel()
@@ -175,7 +176,9 @@ export default class SuggestControl<V, O extends SuggestOptionModel<V>> extends 
     this.cancel()
   }
 
-  private onShowClick: () => void = () => {
+  private onShowClick: React.MouseEventHandler = (event) => {
+    event.preventDefault()
+    event.stopPropagation()
     this.setState({
       focused: true,
     })