+++ packages/core/src/transition/SimpleTransitionProps.ts
@@ -1,6 +1,6 @@
 import {TransitionProps} from 'react-transition-group/Transition'
 
-export default interface SimpleTransitionProps extends TransitionProps {
+type SimpleTransitionProps = TransitionProps & {
   enterClassName?: (timeout: number) => string
   enteredClassName?: (timeout: number) => string
   enteringClassName?: (timeout: number) => string
@@ -8,3 +8,5 @@ export default interface SimpleTransitionProps extends TransitionProps {
   exitingClassName?: (timeout: number) => string
   exitedClassName?: (timeout: number) => string
 }
+
+export default SimpleTransitionProps