+++ packages/desktop/src/suggest-field/SuggestField.tsx
@@ -21,7 +21,7 @@ import {MenuItem} from '../menu'
 import SuggestFieldProps from './SuggestFieldProps'
 import SuggestFieldOptionsModel from './SuggestFieldOptionModel'
 
-const CardPos = Card.withComponent(Pos)
+const CardPos = styled(Card)().withComponent(Pos)
 const CardItem = styled(Card)().withComponent(MenuItem)
 
 const dropDownContainerRef: RefObject<HTMLDivElement> = createRef()
@@ -71,7 +71,7 @@ export const SuggestField = <V extends {}>({
                 <InputField
                   title={props.title}
                   active={renderProps.focused || !!props.suggest || !!props.placeholder}
-                  input={
+                  input={(
                     <BasicInput
                       ref={renderProps.inputRef}
                       name={props.name}
@@ -95,7 +95,7 @@ export const SuggestField = <V extends {}>({
                         renderProps.onItemKeyDown
                       )}
                     />
-                  }
+                  )}
                   hint={props.hint}
                   error={props.error}
                   help={props.help}
@@ -186,3 +186,5 @@ export const SuggestField = <V extends {}>({
 SuggestField.defaultProps = {
   equals: (a: any, b: any) => a === b,
 }
+
+SuggestField.displayName = 'SuggestField'